+++ warehouse/src/main/java/com/scholarscore/api/persistence/goalCalculators/BehaviorGoalCalc.java
@@ -1,14 +1,10 @@
 package com.scholarscore.api.persistence.goalCalculators;
 
 import com.scholarscore.api.persistence.mysql.BehaviorPersistence;
-import com.scholarscore.api.persistence.mysql.jdbc.BehaviorJdbc;
 import com.scholarscore.models.Behavior;
 import com.scholarscore.models.BehaviorGoal;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
 import java.util.Collection;
-import java.util.Collections;
 import java.util.stream.Collectors;
 
 /**
@@ -27,13 +23,13 @@ public class BehaviorGoalCalc implements GoalCalc<BehaviorGoal> {
         this.behaviorPersistence = behaviorPersistence;
     }
 
-    public Float calculateGoal(BehaviorGoal goal) {
+    public Double calculateGoal(BehaviorGoal goal) {
         Collection<Behavior> studentBehaviors = behaviorPersistence.selectAll(goal.getStudent().getId());
         //Make sure behavior category matches and it is within the dates specified.
         Collection<Behavior> relevantBehaviors = studentBehaviors.stream().filter(bg -> bg.getBehaviorCategory().equals(goal.getBehaviorCategory()))
                 .filter(bg -> bg.getBehaviorDate().after(goal.getStartDate()) && bg.getBehaviorDate().before(goal.getEndDate()))
                 .collect(Collectors.toList());
 
-        return new Float(relevantBehaviors.size());
+        return new Double(relevantBehaviors.size());
     }
 }
